requirements.txt
pandas
requests
matplotlib
Pillow
pytz
python-dotenv
playwright
